Added configuration option to disable the use of Discord - This will remove the /chat route - Chat will not be supported without Discord
