+++ templates/client.mustache
@@ -111,9 +111,12 @@
 						</div>
 					</div>
 				</div>
+
+				{{#client_id}}
 				<div id="chat">
 					<iframe src="/chat" frameborder="0"></iframe>
 				</div>
+				{{/client_id}}
 				<div id="notifications">
 					<div id="notification-template" class="notification">
 						<div class="notification-content">The maximum amount of marbles has been reached! No more marbles can be entered.</div>
